+++ com.unity.netcode.gameobjects/Runtime/NetworkVariable/Collections/NetworkList.cs
@@ -429,7 +429,7 @@ namespace Unity.Netcode
         public void Add(T item)
         {
             // check write permissions
-            if (CannotWrite)
+            if (CannotWrite())
             {
                 LogWritePermissionError();
                 return;
@@ -456,7 +456,7 @@ namespace Unity.Netcode
         public void Clear()
         {
             // check write permissions
-            if (CannotWrite)
+            if (CannotWrite())
             {
                 LogWritePermissionError();
                 return;
@@ -494,7 +494,7 @@ namespace Unity.Netcode
         public bool Remove(T item)
         {
             // check write permissions
-            if (CannotWrite)
+            if (CannotWrite())
             {
                 LogWritePermissionError();
                 return false;
@@ -543,7 +543,7 @@ namespace Unity.Netcode
         public void Insert(int index, T item)
         {
             // check write permissions
-            if (CannotWrite)
+            if (CannotWrite())
             {
                 LogWritePermissionError();
                 return;
@@ -579,7 +579,7 @@ namespace Unity.Netcode
         public void RemoveAt(int index)
         {
             // check write permissions
-            if (CannotWrite)
+            if (CannotWrite())
             {
                 throw new InvalidOperationException("Client is not allowed to write to this NetworkList");
             }
@@ -613,7 +613,7 @@ namespace Unity.Netcode
         public void Set(int index, T value, bool forceUpdate = false)
         {
             // check write permissions
-            if (CannotWrite)
+            if (CannotWrite())
             {
                 LogWritePermissionError();
                 return;